requirements.txt
Django>=5.0,<6.0 
djangorestframework>=3.16,<3.17
djangorestframework_simplejwt>=5.5,<5.6
requests>=2.32,<2.33
psycopg2-binary>=2.9,<2.10
drf-yasg>1.21,<1.22